document-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.17.6
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.6
python-multipart>=0.0.5
python-jose>=3.3.0
passlib>=1.7.4
pydantic>=2.0.0
pydantic-settings>=2.0.0
boto3>=1.28.0
aiofiles>=0.8.0
httpx>=0.25.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-dotenv>=1.0.0
pytest>=7.3.0
python-magic>=0.4.27
requests>=2.28.0
minio>=7.1.0
fakeredis
motor>=3.0.0

# LLM Adapters (choose one or more)
# openai
# anthropic
# Add other LLM libraries as needed
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pypdf2
python-docx 